refactor(shared): SCRUM-119 Move RateLimiter classes to dedicated directory
